backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
qdrant-client
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
python-dotenv
markdown
beautifulsoup4

